Update README.md for tutorial links
Currently, the section which contains the list of various Colab tutorials and resources contains incomplete links, and ... as one of the bullet points. 

I have updated these links to include all the Colab tutorial links from the documentation page, and removed that ... which makes it easier to find specific resources, and links.